Move getAssetPath function to locals middleware
